Change to python 3
Change to python 3

+++ functions.py
@@ -24,7 +24,7 @@ def start_game():
         win.border(0)
         win.addstr(0, 2, 'Score : ' + str(score) + ' ')                # Printing 'Score' and
         win.addstr(0, 27, ' SNAKE ')                                   # 'SNAKE' strings
-        win.timeout(150 - (len(snake)/5 + len(snake)/10)%120)          # Increases the speed of Snake as its length increases
+        win.timeout(150 - (int(len(snake)/5) + int(len(snake)/10))%120)          # Increases the speed of Snake as its length increases
         
         prevKey = key                                                  # Previous key pressed
         event = win.getch()
